requirements.txt
# Core ML/DL
tensorflow==2.16.2
keras==3.0.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
scipy==1.13.1
imbalanced-learn==0.12.4
optuna==3.6.1
joblib==1.4.2

# Data IO
pyarrow==17.0.0
fastparquet==2024.5.0

# Visualization
matplotlib==3.9.0
seaborn==0.13.2

# Config / utils
PyYAML==6.0.1
python-dotenv==1.0.1
mlflow==2.10.0
boto3==1.34.144

# Optional API (future)
# fastapi==0.115.0
# uvicorn==0.32.0
# pydantic==2.10.0
